src/scheduler_fapp/requirements.txt
# Azure Functions – Python worker
azure-functions==1.20.0

# Durable Functions (bindings + Python SDK)
azure-functions-durable==1.3.2   

# Anything your activities/orchestrations might need
requests>=2.32.0
